requirements.txt
# ----- PyTorch（使用你實測可用的版本）-----
torch==2.9.1+cpu
torchvision==0.24.1+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# ----- Streamlit -----
streamlit>=1.30

# ----- Data Processing -----
pandas>=1.5
numpy>=1.24
scikit-learn>=1.2

# ----- Image / NLP / Utils -----
wordcloud>=1.9
Pillow>=9.5
ftfy>=6.1.1
regex>=2023.6.3
tqdm>=4.65.0

# ----- HTTP -----
requests>=2.31

# ----- OpenAI CLIP -----
git+https://github.com/openai/CLIP.git
